Worker threads may not throttle down properly.
If worker threads are executing the ThreadProc (the function to execute), then they will miss broadcasts to wake up (since they're already awake).

Pre-test if we need to throttle down prior to waiting for condition.  This insures that the maximum worker thread count is always honored, allowing the ThrottleDown unit test to will always succeed.