Fix problem identifying Red Hat systems properly in SCX_OperatingSystem
Since 2016 TP5 kits can be used in 2012R2, the caption for the
SCX_OperatingSystem class has to be compatible with the 2012R2 MPs.
This change allows the caption to be compatible by fetching the
OSFullName parameter from the /etc/redhat-release file (which is
what older 2012R2 kits did).

Added new unit tests for Red Hat 7.0 to verify functionality with
both /etc/redhat-release and /etc/os-release.