No longer reset network interface counters on AIX
Network interface stats can be retrieved with a command like:
    netstat -v | grep -i elap

On AIX, when enumerating network statistics via omicli like:
   omicli ei root/scx SCX_EthernetPortStatistics
then the network statistics were being reset. This was due a a
deliberate ioctl to reset the network stats. While there was a
comment to that effect, there was no explanation as to WHY this
was being done.

Testing has indicated no unanticipated consequences to this change,
so we're committing it.