Fixing Bug#6362479 - dmSlave traversal
GetDMSlaves results in an erroneous output if the slave of a dm-device is also a dm-device. In such cases, we must traverse to the slaves till we reach a physical disk.
Two tests are introduced to test this functionality to test positive and negative scenarios.
Also included is a minor fix in diskdepend.cpp, which results in empty name under some rare circumstances.