Add log entries for enumeration of network provider
Add log entries for enumeration of network provider. This was
needed to identify very slow network enumeration.

It turns out that the slow network provider enumeration was due to:

1) Network speed detection (bind() delay), and
2) MAC network address detection (getkerninfo() delay)

While the problem was ultimately a network configuration issue on a
customer's AIX system, these changes are being committed to allow for
easier detection in the future, should this arise again.