WIP: Fix system-specific builds to be properly supported
The intent is that, when configuring:

if --enable-microsoft is specified without --enable-system-build
   (and without --enable-ulinux),

then build a kit (assuming --enable-native-kits) that is built for
the current system only. This is significantly faster for development
since we don't need to build three sets of images for the various
flavors of OpenSSL.

Unfortunately, this was broken in December 2015 with the redefinition
of "ULINUX" to be a universal kit.

With this changeset, we're close. Unfortunately, PF_DISTRO needs to
be set to something like REDHAT or SUSE or UBUNTU, and it's not done
today. Look at changes to pal/build/configure (this file) back around
December, 2015, and bring back the code to set PF_DISTRO appropriately
*IF* --enable-ulinux was not specified.