Enabled testBootTime and testUpTime testcases
testBootTime and testUpTime test cases were failing because boot time read from
/etc/utmpx file was incorrect. It was found that program writing the boot time
in /etc/utmpx is a 32bit process and struct utmpx size was 140 bytes because
time_t tv_sec in 32bit process is a 32bit signed integer. In case of 64bit process time_t
corresponds to 64bit integer so sizeof struct utmpx is 152 bytes due to structure
padding. So when tv_sec was referenced in the struct utmpx then it was referring
to another variable corresponding to a 32bit struct whose value was 0.

This change declares a new struct utmpx32bits which defines tv_sec as 32bit signed
integer. Now sizeof struct utmpx32bits is 140 in 64bit program similar to sizeof
struct utmpx in 32bit program.