Fixed ioctl fail issue in AIX 7.1
IBM APAR: APAR IJ15829
Case Details: https://www.ibm.com/mysupport/s/case/5000z00001J44vnAAB/ioctl-fails-on-aix-machine-with-shient-driver

Root cause of the ioctl error has been described as a failure within ioctl for NDD_GET_ALL_STATS.
That will in turn call: shient_all_stats_t which you are using in this case with a size of:  "Its size is only 8192 bytes."

The problem is that within the driver code calls from NDD_GET_ALL_STATS are
expected to pass thru a larger value: "that is the size of entcore_all_stats_t + hw_stats_size:
- There is no way that can be known. The code is not exposed to the outside world.

Thus, when the ioctl passes thru in the custom code provide, the value of entcore_all_stats_t is passed
but does not include the allocation for "hw_stats_size" and thus the ioctl fails.

This patch is only work around as suggested by IBM to increase size of the union to 50kb to provide
additional space for the "hw_stats_size".