iconv_open fixes.
 iconv_open(3) says:

A conversion descriptor contains a conversion state. After creation
using iconv_open(), the state is in the initial state. Using iconv(3)
modifies the descriptor's conversion state. (This implies that a con-
version descriptor can not be used in multiple threads simultaneously.)
To bring the state back to the initial state, use iconv(3) with NULL as
inbuf argument.